backend/requirements.txt
alembic==1.13.1
async-timeout==4.0.3
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
contextlib2==21.6.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.3
schema==0.7.5
SQLAlchemy==2.0.29
termcolor==2.4.0
typing_extensions==4.10.0
Werkzeug==3.0.1
zipp==3.18.1
